+++ clojure-mode.el
@@ -69,7 +69,6 @@
   (defvar paredit-mode))
 
 (require 'cl)
-(require 'inf-lisp)
 (require 'imenu)
 
 (declare-function lisp-fill-paragraph  "lisp-mode" (&optional justify))
@@ -98,21 +97,6 @@
   "Face used to font-lock interop method names (camelCase)."
   :group 'clojure
   :package-version '(clojure-mode . "3.0.0"))
-
-(defcustom clojure-load-command  "(clojure.core/load-file \"%s\")\n"
-  "Format-string for building a Clojure expression to load a file.
-This format string should use `%s' to substitute a file name and
-should result in a Clojure expression that will command the
-inferior Clojure to load that file."
-  :type 'string
-  :group 'clojure
-  :safe 'stringp)
-
-(defcustom clojure-inf-lisp-command "lein repl"
-  "The command used by `inferior-lisp-program'."
-  :type 'string
-  :group 'clojure
-  :safe 'stringp)
 
 (defcustom clojure-defun-style-default-indent nil
   "When non-nil, use default indenting for functions and macros.
@@ -162,24 +146,9 @@ For example, \[ is allowed in :db/id[:db.part/user]."
 
 (defvar clojure-mode-map
   (let ((map (make-sparse-keymap)))
-    (set-keymap-parent map lisp-mode-shared-map)
-    (define-key map (kbd "C-M-x")   'lisp-eval-defun)
-    (define-key map (kbd "C-x C-e") 'lisp-eval-last-sexp)
-    (define-key map (kbd "C-c C-e") 'lisp-eval-last-sexp)
-    (define-key map (kbd "C-c C-l") 'clojure-load-file)
-    (define-key map (kbd "C-c C-r") 'lisp-eval-region)
-    (define-key map (kbd "C-c C-z") 'clojure-display-inferior-lisp-buffer)
     (define-key map (kbd "C-:") 'clojure-toggle-keyword-string)
     (easy-menu-define clojure-mode-menu map "Clojure Mode Menu"
       '("Clojure"
-        ["Eval Top-Level Expression" lisp-eval-defun]
-        ["Eval Last Expression" lisp-eval-last-sexp]
-        ["Eval Region" lisp-eval-region]
-        "--"
-        ["Run Inferior Lisp" clojure-display-inferior-lisp-buffer]
-        ["Display Inferior Lisp Buffer" clojure-display-inferior-lisp-buffer]
-        ["Load File" clojure-load-file]
-        "--"
         ["Toggle between string & keyword" clojure-toggle-keyword-string]
         ["Mark string" clojure-mark-string]
         ["Insert ns form at point" clojure-insert-ns-form-at-point]
@@ -202,11 +171,6 @@ For example, \[ is allowed in :db/id[:db.part/user]."
     ;; Make hash a usual word character
     (modify-syntax-entry ?# "_ p" table)
     table))
-
-(defvar clojure-prev-l/c-dir/file nil
-  "Record last directory and file used in loading or compiling.
-This holds a cons cell of the form `(DIRECTORY . FILE)'
-describing the last `clojure-load-file' or `clojure-compile-file' command.")
 
 (defconst clojure-mode-version "3.1.0-snapshot"
   "The current version of `clojure-mode'.")
@@ -288,7 +252,6 @@ ENDP and DELIMITER."
   (setq-local indent-line-function 'clojure-indent-line)
   (setq-local lisp-indent-function 'clojure-indent-function)
   (setq-local lisp-doc-string-elt-property 'clojure-doc-string-elt)
-  (setq-local inferior-lisp-program clojure-inf-lisp-command)
   (setq-local parse-sexp-ignore-comments t)
   (setq-local prettify-symbols-alist clojure--prettify-symbols-alist)
   (setq-local open-paren-in-column-0-is-defun-start nil))
@@ -351,25 +314,6 @@ If JUSTIFY is non-nil, justify as well as fill the paragraph."
                            fill-column))
             (fill-prefix (clojure-adaptive-fill-function)))
         (do-auto-fill)))))
-
-(defun clojure-display-inferior-lisp-buffer ()
-  "Display a buffer bound to `inferior-lisp-buffer'."
-  (interactive)
-  (if (and inferior-lisp-buffer (get-buffer inferior-lisp-buffer))
-      (pop-to-buffer inferior-lisp-buffer t)
-    (run-lisp inferior-lisp-program)))
-
-(defun clojure-load-file (file-name)
-  "Load a Clojure file FILE-NAME into the inferior Clojure process."
-  (interactive (comint-get-source "Load Clojure file: "
-                                  clojure-prev-l/c-dir/file
-                                  '(clojure-mode) t))
-  (comint-check-source file-name) ; Check to see if buffer needs saved.
-  (setq clojure-prev-l/c-dir/file (cons (file-name-directory file-name)
-                                        (file-name-nondirectory file-name)))
-  (comint-send-string (inferior-lisp-proc)
-                      (format clojure-load-command file-name))
-  (switch-to-lisp t))